Change directory.
Moved mentioned files to a parent directory 'com.askie01' as it's project's parent directory.
