+++ src/main/java/com/askie01/filters/CacheControlFilter.java
@@ -1,0 +1,28 @@
+package com.askie01.filters;
+
+import jakarta.servlet.*;
+import jakarta.servlet.annotation.WebFilter;
+import jakarta.servlet.http.HttpServletRequest;
+import jakarta.servlet.http.HttpServletResponse;
+import lombok.SneakyThrows;
+
+@WebFilter({"/user_page.jsp", "/list_products.jsp", "/explore_products.jsp", "/edit_account.jsp"})
+public class CacheControlFilter implements Filter {
+
+    @Override
+    @SneakyThrows
+    public void doFilter(ServletRequest request, ServletResponse response, FilterChain chain) {
+        final HttpServletResponse httpResponse = (HttpServletResponse) response;
+        final HttpServletRequest httpRequest = (HttpServletRequest) request;
+        final Object user = httpRequest.getSession().getAttribute("user");
+
+        if (user == null) {
+            httpResponse.sendRedirect("login_page.jsp");
+        }
+
+        httpResponse.setHeader("Cache-Control", "no-cache, no-store, must-revalidate");
+        httpResponse.setHeader("Pragma", "no-cache");
+        httpResponse.setDateHeader("Expires", 0);
+        chain.doFilter(request, response);
+    }
+}
