+++ src/main/java/com/askie01/filters/RegistrationLoginDataFilter.java
@@ -1,0 +1,32 @@
+package com.askie01.filters;
+
+import jakarta.servlet.*;
+import jakarta.servlet.annotation.WebFilter;
+import jakarta.servlet.http.HttpServletResponse;
+import lombok.SneakyThrows;
+import org.apache.logging.log4j.LogManager;
+import org.apache.logging.log4j.Logger;
+import com.askie01.validators.UserValidator;
+
+@WebFilter("/register")
+public class RegistrationLoginDataFilter implements Filter {
+    private final Logger log = LogManager.getLogger(RegistrationLoginDataFilter.class.getName());
+
+    @Override
+    @SneakyThrows
+    public void doFilter(ServletRequest request, ServletResponse response, FilterChain chain) {
+        final UserValidator userValidator = new UserValidator();
+
+        final String login = request.getParameter("login");
+        final String password = request.getParameter("password");
+
+        if (userValidator.isValidLoginData(login, password)) {
+            log.info("Received valid login data. \nLogin: '{}' \npassword: '{}'", login, password);
+            chain.doFilter(request, response);
+        } else {
+            log.warn("Received invalid login data. \nLogin: '{}' \npassword: '{}'", login, password);
+            final HttpServletResponse httpResponse = (HttpServletResponse) response;
+            httpResponse.sendRedirect("registration_page.jsp");
+        }
+    }
+}
