+++ src/main/java/com/askie01/products/ProductCategory.java
@@ -1,0 +1,5 @@
+package com.askie01.products;
+
+public enum ProductCategory {
+    ELECTRONICS, BOOKS
+}
