+++ src/main/java/com/askie01/services/product/ExploreProductsServlet.java
@@ -1,0 +1,17 @@
+package com.askie01.services.product;
+
+import jakarta.servlet.annotation.WebServlet;
+import jakarta.servlet.http.HttpServlet;
+import jakarta.servlet.http.HttpServletRequest;
+import jakarta.servlet.http.HttpServletResponse;
+import lombok.SneakyThrows;
+
+@WebServlet("/user/explore_products")
+public class ExploreProductsServlet extends HttpServlet {
+
+    @Override
+    @SneakyThrows
+    protected void doPost(HttpServletRequest req, HttpServletResponse resp) {
+
+    }
+}
