+++ src/main/java/com/askie01/services/product/ListProductsServlet.java
@@ -1,0 +1,16 @@
+package com.askie01.services.product;
+
+import jakarta.servlet.annotation.WebServlet;
+import jakarta.servlet.http.HttpServlet;
+import jakarta.servlet.http.HttpServletRequest;
+import jakarta.servlet.http.HttpServletResponse;
+import lombok.SneakyThrows;
+
+@WebServlet("/user/list_products")
+public class ListProductsServlet extends HttpServlet {
+    @Override
+    @SneakyThrows
+    protected void doPost(HttpServletRequest request, HttpServletResponse response) {
+
+    }
+}
