+++ src/main/java/com/askie01/services/user/LogoutServlet.java
@@ -1,0 +1,23 @@
+package com.askie01.services.user;
+
+import jakarta.servlet.annotation.WebServlet;
+import jakarta.servlet.http.HttpServlet;
+import jakarta.servlet.http.HttpServletRequest;
+import jakarta.servlet.http.HttpServletResponse;
+import jakarta.servlet.http.HttpSession;
+import lombok.SneakyThrows;
+
+@WebServlet("/logout")
+public class LogoutServlet extends HttpServlet {
+    @SneakyThrows
+    protected void doPost(HttpServletRequest request, HttpServletResponse response) {
+        final HttpSession session = request.getSession();
+        session.invalidate();
+        response.sendRedirect("welcome_page.jsp");
+    }
+
+    @SneakyThrows
+    protected void doGet(HttpServletRequest request, HttpServletResponse response) {
+        response.sendRedirect("login_page.jsp");
+    }
+}
