+++ src/main/java/com/askie01/validators/DataValidatorLogger.java
@@ -1,0 +1,17 @@
+package com.askie01.validators;
+
+import lombok.RequiredArgsConstructor;
+import org.apache.logging.log4j.Logger;
+
+@RequiredArgsConstructor
+public class DataValidatorLogger {
+    private final Logger log;
+
+    public void logDataValidation(String data, boolean isValid) {
+        if (isValid) {
+            log.info("Valid data: '{}'", data);
+        } else {
+            log.warn("Invalid data: '{}'", data);
+        }
+    }
+}
