--- src/main/java/products/ProductCategory.java
@@ -1,5 +1,0 @@
-package products;
-
-public enum ProductCategory {
-    ELECTRONICS, BOOKS
-}
