--- src/main/java/services/product/ExploreProductsServlet.java
@@ -1,17 +1,0 @@
-package services.product;
-
-import jakarta.servlet.annotation.WebServlet;
-import jakarta.servlet.http.HttpServlet;
-import jakarta.servlet.http.HttpServletRequest;
-import jakarta.servlet.http.HttpServletResponse;
-import lombok.SneakyThrows;
-
-@WebServlet("/user/explore_products")
-public class ExploreProductsServlet extends HttpServlet {
-
-    @Override
-    @SneakyThrows
-    protected void doPost(HttpServletRequest req, HttpServletResponse resp) {
-
-    }
-}
